FIX: 이동 관련 버그 수정

+++ Client/SocketCore/SocketEvent.h
@@ -11,8 +11,8 @@ struct MovingEvent
 {
     
     uint32_t userId;
-    int32_t locationX;
-    int32_t locationY;
+    float locationX;
+    float locationY;
 };
 
 struct BroadcastingEnterEvent
